SENA-119 Implement the Cash Closing Query - Backend

+++ LaCiguenia-Backend/src/main/java/LaCiguenia/commons/domains/dto/opening/OpeningDTO.java
@@ -15,5 +15,5 @@ public class OpeningDTO implements Serializable {
     private Integer openingId;
     private LocalDate openingDate;
     private String openingStore;
-    private Integer openingTotal;
+    private Double openingTotal;
 }
